+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 1.12.0 (?)
 Changes in this release:
+- Add option to specify project and environment names
 
 # v 1.11.0 (2023-02-20)
 Changes in this release: